+++ qmq-delay-server/src/main/java/qunar/tc/qmq/delay/store/log/DelaySegment.java
@@ -38,7 +38,7 @@ public interface DelaySegment<T> {
 
     long getFlushedPosition();
 
-    int getSegmentBaseOffset();
+    long getSegmentBaseOffset();
 
     long validate() throws IOException;
